Initial commit.
This is a new repo for my research code. It is meant to be a filtered version of the stuff I wrote over the last one and a half years. In particular, it contains a simple python version of the eye model, with the new fitting procedure being implemented both in python and in cpp.