Expose 3d sphere fit rms residual but disable it by default
Currently, the rms residual is always calculated but not used anywhere. To save performance, we disable this calculation by default. It can be enabled anywhere in the calculation chain:
- `Detector3D()`
- `TwoSphereModel.estimate_sphere_center()`
- `TwoSphereModel.estimate_sphere_center_3d()`

`TwoSphereModel.estimate_sphere_center()` returns a `SphereCenterEstimates` named tuple which was extended with a `rms_residual` field. This field is filled with a float if the function was called with `calculate_rms_residual=True`, else it is either filled with `None` or `NaN`.