Use long-term-model pupil-circle normal for gaze direction when frozen
Previously, the resulting 3d pupil circle was always composed based on:
- short-term normal
- long-term center
- long-term radius

The short-term model is intended as slippage compensation for gaze directions. If the eye is still the resulting short-term model is very noisy and therefore the gaze signal is too.

With this PR, we change the 3d pupil circle composition to use the long-term model for the circle normal, too, if the long-term model is frozen. Freezing the model already implies that there will be no slippage. Therefore, we can rely on the long-term model for gaze direction, too.