Set model_confidence to 0.1 if parameter out of physiological range
In the current version, model_confidence was fixed to 1.0 as there was no way to estimate it. Starting with this PR, model_confidence is 1.0 by default but set to
0.1 if at least one model output exceeds its physiologically reasonable range. If the ranges are exceeded, it is likely that the model is either not fit
well or the 2d input ellipse was a false detection.

- phi: [-80, 80] degrees
- theta: [-80, 80] degrees
- pupil diameter: [1.0, 9.0] mm
- eyeball center x: [-10, 10] mm
- eyeball center y: [-10, 10] mm
- eyeball center z: [20, 75] mm

The model_confidence will be set to 0.0 if the gaze direction could not be calculated.